Added reservation_creator & test file, migrated all code & tests into new files.
