+++ test/test_helper.rb
@@ -18,3 +18,4 @@ require_relative '../lib/room'
 require_relative '../lib/reservation_manager'
 require_relative '../lib/calendar'
 require_relative '../lib/block'
+require_relative '../lib/reservation_creator'
